phase-1_research-automation/requirements.txt
google-generativeai>=0.8.3
google-genai>=0.2.0
python-dotenv>=1.0.0
jinja2>=3.1.0
supabase>=2.0.0
